# Speed up alert fingerprint generation