--- a/models/alert_test.go
+++ b/models/alert_test.go
@@ -2,6 +2,7 @@ package models_test
 
 import (
 	"testing"
+	"time"
 
 	"github.com/cloudflare/unsee/models"
 )
@@ -50,5 +51,51 @@ func TestAlertState(t *testing.T) {
 			t.Errorf("alert.IsSilenced() returned %t while %t was expected for alert %v",
 				testCase.alert.IsSilenced(), testCase.isSilenced, testCase.alert)
 		}
+	}
+}
+
+func BenchmarkLabelsFingerprint(b *testing.B) {
+	alert := models.Alert{
+		Labels: map[string]string{
+			"foo1":        "bar1",
+			"foo1bar1":    "545jjjssd",
+			"foo1xxxx":    "bdjjs88ff",
+			"agdfdfd":     "bar1",
+			"fossdsf3o1":  "bar11111",
+			"fdfdgfdgoo1": "bar1",
+		},
+	}
+	for n := 0; n < b.N; n++ {
+		alert.LabelsFingerprint()
+	}
+}
+
+func BenchmarkLabelsContent(b *testing.B) {
+	alert := models.Alert{
+		Annotations: map[string]string{
+			"foo": "bar",
+			"abc": "Neque porro quisquam est qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit...",
+		},
+		Labels: map[string]string{
+			"foo1":        "bar1",
+			"foo1bar1":    "545jjjssd",
+			"foo1xxxx":    "bdjjs88ff",
+			"agdfdfd":     "bar1",
+			"fossdsf3o1":  "bar11111",
+			"fdfdgfdgoo1": "bar1",
+		},
+		State:    models.AlertStateActive,
+		StartsAt: time.Date(2015, time.March, 10, 0, 0, 0, 0, time.UTC),
+		Alertmanager: []models.AlertmanagerInstance{
+			models.AlertmanagerInstance{
+				Name:  "default",
+				URI:   "http://localhost",
+				State: models.AlertStateActive,
+			},
+		},
+	}
+	alert.UpdateFingerprints()
+	for n := 0; n < b.N; n++ {
+		alert.LabelsFingerprint()
 	}
 }